runtime/requirements.txt
boto3<2.0.0
botocore~=1.23.11
chalice==1.26.2
pyjwt==2.3.0
